Shared kernelconnection wrapper
